adding roles function in users

+++ journalApp/src/main/java/com/BackendX/journalApp/Service/JournalEntryService.java
@@ -1,11 +1,15 @@
 package com.BackendX.journalApp.Service;
 
+import ch.qos.logback.core.encoder.EchoEncoder;
 import com.BackendX.journalApp.entity.JournalEntry;
+import com.BackendX.journalApp.entity.User;
 import com.BackendX.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,10 +17,30 @@ import java.util.Optional;
 public class JournalEntryService {
     @Autowired
     private JournalEntryRepository journalEntryRepository;
+    @Autowired
+    private UserService userService;
 
-    public void saveEntry(JournalEntry journalEntry){
-        journalEntryRepository.save(journalEntry);
+    @Transactional
+    public void saveEntry(JournalEntry journalEntry, String userName){
+        try {
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveUser(user);
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("an error ocured while saving the entry.",e);
+        }
+
+
+
     }
+    public void saveEntry(JournalEntry journalEntry){
+       journalEntryRepository.save(journalEntry);
+
+    }
+
     public List<JournalEntry> getAll(){
         return journalEntryRepository.findAll();
     }
@@ -24,8 +48,26 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id){
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry",e);
+        }
+        return removed;
+
+
+
     }
+
 
 }
